순수 자바로 오브젝트 팩토리까지 구현
