--- src/main/java/com/dasom/tobi/dao/DConnectionMakeer.java
@@ -1,11 +1,0 @@
-package com.dasom.tobi.dao;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-
-public class DConnectionMakeer implements ConnectionMaker{
-    @Override
-    public Connection makeConnection() throws ClassNotFoundException, SQLException {
-        return null;
-    }
-}
